Clean up: Remove absolute paths in config's; rename toptagging->tagging and weaver->miniweaver (and change imports accordingly); fix experiments/eventgen/wrappers.py import from tagging (forgot to do this earlier)
