+++ config/default_tagging.yaml
@@ -1,0 +1,33 @@
+data:
+ beam_reference: spacelike
+ two_beams: true
+ beam_token: true
+ add_time_reference: false
+ add_pt: false
+ reinsert_channels: false
+ rescale_data: true
+
+train: true
+training:
+ iterations: 10
+ batchsize: 128
+ lr: 1e-3
+ validate_every_n_steps: 5000
+
+ force_xformers: false
+
+evaluation:
+ batchsize: 128
+ save_roc: false
+ eval_set: [test]
+
+plotting:
+ roc: true
+ loss: true
+ score: true
+
+defaults:
+ - model: gatr_tagging
+ - default
+ - hydra
+ - _self_
