+++ config_paper/toptagging.yaml
@@ -1,9 +1,9 @@
 exp_type: toptagging
-exp_name: topt_local_debug
-base_dir: /media/jspinner/shared/Studium/project3/lorentz-gatr
+exp_name: topt_local
+base_dir: .
 
 data:
- data_dir : /media/jspinner/shared/Studium/project3/data
+ data_dir : data/
  dataset: full
  beam_reference: spacelike
  two_beams: true
