+++ config_paper/ttbar.yaml
@@ -1,14 +1,13 @@
 exp_type: ttbar
-exp_name: ttbar_local_debug
-base_dir: /media/jspinner/shared/Studium/project3/lorentz-gatr
+exp_name: ttbar_local
 
 data:
  n_jets: [0]
- data_path_0j: /media/jspinner/shared/Studium/project3/data/ttbar_0j.npy
- data_path_1j: /media/jspinner/shared/Studium/project3/data/ttbar_1j.npy
- data_path_2j: /media/jspinner/shared/Studium/project3/data/ttbar_2j.npy
- data_path_3j: /media/jspinner/shared/Studium/project3/data/ttbar_3j.npy
- data_path_4j: /media/jspinner/shared/Studium/project3/data/ttbar_4j.npy
+ data_path_0j: data/ttbar_0j.npy
+ data_path_1j: data/ttbar_1j.npy
+ data_path_2j: data/ttbar_2j.npy
+ data_path_3j: data/ttbar_3j.npy
+ data_path_4j: data/ttbar_4j.npy
  subsample: null
  train_test_val: [.5, .4, .01]
  base_type: 2
