+++ experiments/eventgen/wrappers.py
@@ -1,11 +1,10 @@
-import math
 import torch
 import numpy as np
 from torch import nn
 
-from gatr.interface import embed_vector, extract_vector, extract_scalar
+from gatr.interface import embed_vector, extract_vector
 from gatr.layers import EquiLinear, GeometricBilinear, ScalarGatedNonlinearity
-from experiments.toptagging.dataset import embed_beam_reference
+from experiments.tagging.embedding import get_spurion
 from experiments.eventgen.cfm import EventCFM
 
 from experiments.eventgen.coordinates import (
@@ -122,8 +121,12 @@ class GAPCFM(EventCFM):
 
         # mv embedding
         mv = embed_vector(x.reshape(x.shape[0], -1, 4))
-        beam = embed_beam_reference(
-            mv, self.beam_reference, self.add_time_reference, self.two_beams
+        beam = get_spurion(
+            self.beam_reference,
+            self.add_time_reference,
+            self.two_beams,
+            device=mv.device,
+            dtype=mv.dtype,
         )
         if beam is not None:
             beam = beam.unsqueeze(0).expand(*mv.shape[:-2], beam.shape[-2], 16)
@@ -204,7 +207,7 @@ class GATrCFM(EventCFM):
         beam_reference : str
             Type of beam reference used to break the Lorentz symmetry
             Options: [None, "xyplane", "spacelike", "lightlike", "timelike"]
-            See experiments.toptagging.dataset.py::embed_beam_reference for details
+            See experiments.toptagging.embedding.py::get_spurion for details
         two_beams : bool
             If beam_reference in ["spacelike", "lightlike", "timelike"],
             decide whether only (alpha,0,0,1) or both (alpha,0,0,+/-1) are included
@@ -263,8 +266,12 @@ class GATrCFM(EventCFM):
 
         # mv embedding
         mv = embed_vector(x).unsqueeze(-2)
-        beam = embed_beam_reference(
-            mv, self.beam_reference, self.add_time_reference, self.two_beams
+        beam = get_spurion(
+            self.beam_reference,
+            self.add_time_reference,
+            self.two_beams,
+            device=mv.device,
+            dtype=mv.dtype,
         )
         if beam is not None:
             beam = (
