+++ experiments/tagging/experiment.py
@@ -8,10 +8,10 @@ from omegaconf import open_dict
 from sklearn.metrics import roc_curve, roc_auc_score, accuracy_score
 
 from experiments.base_experiment import BaseExperiment
-from experiments.toptagging.dataset import TopTaggingDataset
-from experiments.toptagging.dataset import QGTaggingDataset
-from experiments.toptagging.plots import plot_mixer
-from experiments.toptagging.embedding import embed_tagging_data_into_ga
+from experiments.tagging.dataset import TopTaggingDataset
+from experiments.tagging.dataset import QGTaggingDataset
+from experiments.tagging.plots import plot_mixer
+from experiments.tagging.embedding import embed_tagging_data_into_ga
 from experiments.logger import LOGGER
 from experiments.mlflow import log_mlflow
